* all : mpeg_audio decoder works now with aout3.

+++ modules/codec/mpeg_audio/generic.c
@@ -2,7 +2,7 @@
  * generic.c: MPEG audio decoder
  *****************************************************************************
  * Copyright (C) 1999-2001 VideoLAN
- * $Id: generic.c,v 1.1 2002/08/04 17:23:42 sam Exp $
+ * $Id: generic.c,v 1.2 2002/08/17 15:35:10 fenrir Exp $
  *
  * Authors: Michel Kaempf <maxx@via.ecp.fr>
  *          Michel Lespinasse <walken@via.ecp.fr>
@@ -26,6 +26,7 @@
 #include <string.h>                                    /* memcpy(), memset() */
 
 #include <vlc/vlc.h>
+#include <vlc/aout.h>
 #include <vlc/decoder.h>
 
 #include "generic.h"
@@ -182,7 +183,7 @@ int adec_SyncFrame( adec_thread_t * p_adec, adec_sync_info_t * p_sync_info )
     return 0;
 }
 
-int adec_DecodeFrame( adec_thread_t * p_adec, s16 * buffer )
+int adec_DecodeFrame( adec_thread_t * p_adec, float * buffer )
 {
     int i_total_bytes_read;
